Changes before meeting with stephan. Prefabs for both the music and the smell managers
Both sound and smell now how prefabs that can be simply placed in the scenes and are automatically connected with the corresponding global manager.
For the smell one must must the sphere and adjust the size to see where the smell starts to work. For the smell put the prefab -has no sphere or so- the global mixer is automatically connected. One must only put in each scene the corresponding sound for each local sound source object.