Added first movement to the XRRig with keyboard (and maybe controller too -still has to be tested-
The movement in the horizontal plane is working with the wasd keys. However the height is not being adjusted as wished for.

The player keeps a constan height over the global coordinated planes and not over the terrain.
The current approach tries to detect the collision of a ray going down with the terrain which apparently is not working.
To be checked how this ray collision detection with teh terrain can be done or if keeping a constant height over the terrain can be
implemented somehow else.