Added movement to charachter
The movement of the character was added using actions. Tkeyboad with
mouse was tested and the base functionality is there. The charachter
moves on the terrain and adaptds the height automaticaly and doestn go
thrugh objects. However the angle of the camera looks weird, this must
still be corrected.
This hasnt been tested on the VRSet and still has to be done.